Fix README headers
This fixes the function headers in the README so that they display as
markdown headers instead of the markdown text. I believe the anchor tags
were causing the issue. Using the header link, which downcases the
string, removes special characters, and replaces whitespace with a dash,
references the correct link.